File "vid.py" was added
This file is able to combine a gif file and an audio file to generate a video file that has the duration of the audio file. 
If the gif includes a transparent background, it will get covered with white background, upon conversion to video format.
Use this file as needed!